+++ app/src/main/java/com/qe/qzin/activities/MainActivity.java
@@ -9,7 +9,6 @@ import android.support.v7.app.ActionBarDrawerToggle;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.Toolbar;
-import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 
@@ -17,7 +16,6 @@ import com.parse.FindCallback;
 import com.parse.ParseException;
 import com.parse.ParseQuery;
 import com.parse.ParseUser;
-import com.parse.SaveCallback;
 import com.qe.qzin.R;
 import com.qe.qzin.adapters.EventsAdapter;
 import com.qe.qzin.listeners.EndlessRecyclerViewScrollListener;
@@ -25,7 +23,6 @@ import com.qe.qzin.models.Event;
 import com.qe.qzin.models.User;
 
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 
 import butterknife.BindView;
@@ -61,8 +58,6 @@ public class MainActivity extends BaseActivity implements NavigationView.OnNavig
 
 
     LinearLayoutManager linearLayoutManager = new LinearLayoutManager(this);
-
-    createEventsForTest();
 
     events = getEventData();
     eventsAdapter = new EventsAdapter(this,events);
@@ -161,50 +156,4 @@ public class MainActivity extends BaseActivity implements NavigationView.OnNavig
     //fetch next records from event table in database
     return null;
   }
-
-  private void createEventsForTest() {
-
-    Event newEvent = new Event();
-    newEvent.setTitle("Modern Italian Feast");
-    newEvent.setLocality("San Francisco, CA");
-    newEvent.setDate(new Date());
-    newEvent.setDescription("Your culinary journey will include an elegant five-course tasting menu inspired by " +
-        "rustic Northern Italian inspired cuisine married with a creative, modern twist.");
-    newEvent.setAmount(45);
-
-      newEvent.saveInBackground(new SaveCallback() {
-        @Override
-        public void done(ParseException e) {
-          if(e == null)
-            Log.d("DEBUG","New Record Added in Events");
-          else
-            e.printStackTrace();
-        }
-
-      });
-
-/*
-    Event newEvent2 = new Event();
-
-    newEvent2.setTitle("Escape to India");
-    newEvent2.setLocality("Fremont, CA");
-    newEvent2.setDate(new Date());
-    newEvent2.setDescription("Spending our childhood in India, we have grown up eating Indian food prepared by " +
-        "our families. We invite you to experience some of those flavors to understand what true home cooked Indian " +
-        "food tastes like as you embrace tit bits of our culture. Through the journey, we promise to give you insight " +
-        "on the history behind each dish to ensure you have an enriching and holistic experience");
-    newEvent2.setAmount(39);
-
-    newEvent2.saveInBackground(new SaveCallback() {
-      @Override
-      public void done(ParseException e) {
-        if(e == null)
-          Log.d("DEBUG", "New Record Added in Events");
-        else
-          e.printStackTrace();
-      }
-    });*/
-
-  }
-
 }
